Make register_opt() expect an oslo config CONF object
The plugin register_opt() function was previously written assuming a
temepst private config object (which is a thin layer on top of oslo's)
was being passed to the plugin. This however was not the case, mostly
because the tempest config object isn't created when register_opt()
is called. This commit changes all the references to the local aliases
that the tempest object would have to use dicts for the group like
oslo.config provides.